backend/pyproject.toml
[project]
name = "sosoft-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.17.2",
    "asyncpg>=0.31.0",
    "email-validator>=2.3.0",
    "fastapi>=0.123.5",
    "geoalchemy2>=0.18.1",
    "httpx>=0.28.1",
    "minio>=7.2.20",
    "phonenumbers>=9.0.19",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "sqlalchemy[asyncio]>=2.0.44",
    "twilio>=9.8.8",
    "uvicorn[standard]>=0.38.0",
    "bcrypt>=4.0.0",
]

[dependency-groups]
dev = [
    "aiosqlite>=0.20.0",
    "factory-boy>=3.3.3",
    "mypy>=1.19.0",
    "psycopg2-binary>=2.9.11",
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.14.7",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.coverage.run]
source = ["app"]
omit = ["app/scripts/*"]
concurrency = ["greenlet", "thread"]

[tool.coverage.report]
fail_under = 80
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]
